requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[argon2]
python-multipart
dash
pandas
plotly
requests
psycopg2-binary
